requirements.txt
selenium
webdriver-manager
pyyaml
python-dotenv
gspread
oauth2client
tenacity
pytest
pytest-reporter-html1

